api(syscall): Make getreadcount return uint64 across kernel/user
Unify prototype with sys_getreadcount(uint64) and prevent truncation.
- kernel/defs.h: int -> uint64
- user/user.h  : int -> uint64
This aligns the public interface and avoids future overflow issues.

+++ kernel/defs.h
@@ -109,7 +109,7 @@ int             getpinfo(uint64);
 int             either_copyout(int user_dst, uint64 dst, void *src, uint64 len);
 int             either_copyin(void *dst, int user_src, uint64 src, uint64 len);
 void            procdump(void);
-int             getreadcount(void);  // Track read() syscall count
+uint64             getreadcount(void);  // Track read() syscall count
 
 // swtch.S
 void            swtch(struct context*, struct context*);
